Moved compressPublicKey and uncompressPublicKey to Key
